middleware/requirements.txt
# Middleware Dependencies - All versions
# Use requirements-mqtt.txt for MQTT only
# Use requirements-signalr.txt for SignalR only

# Common dependencies
pydantic==2.5.0
structlog==23.2.0
pyyaml==6.0.1

# Protocol-specific dependencies (install as needed)
aiomqtt==2.0.0          # For MQTT version
signalrcore==0.9.3      # For SignalR version
